+++ CHANGELOG.md
@@ -1028,6 +1028,7 @@
 # 0.99.73
 
 * Add GZIPping capability to data bus storage - can be enabled by attaching `.UseCompression()` in the data bus configuration builder
+* Factor forwarding of failed messages to error queues out into `PoisonQueueErrorHandler` which implements `IErrorHandler`. Make room for customizing what to do about failed messages.
 
 ---